requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-utils==0.7.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.5
packaging==24.1
pluggy==1.5.0
psutil==5.9.8
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyMySQL==1.1.1
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
testing.common.database==2.0.3
testing.mysqld==1.4.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
